consolidated test into iPgmUnit
